requirements.txt
discord.py
python-dotenv
Pillow
pyodbc
topggpy
requests